Refactor: BoardStatus를 ChessView로 변경 BoardStatus에서 Board의 점수를 구하거나 하는 것들을 Game 클래스로 이동시키고, ChessView로 변경하여 출력하는 것만 담당하도록 했다.
